Consolidate zero-copy primitives in moirai-core, enforce SSOT
