+++ CHECKLIST.md
@@ -115,7 +115,9 @@
 - [ ] Comprehensive test suite
 - [ ] Performance benchmarks
 - [ ] Example applications
-- [ ] Documentation improvements
+- [x] Documentation improvements (SSOT and consolidation notes)
 - [ ] API stabilization
 - [ ] Performance profiling
 - [ ] Production readiness review
+- [x] SSOT consolidation: zero-copy moved to `moirai_core::communication::zero_copy`
+- [x] Iterator windows/chunks consolidated under `moirai_iter::windows`
