--- moirai-transport/src/zero_copy.rs
@@ -1,1169 +1,0 @@
-//! Zero-copy communication channels.
-//!
-//! This module provides high-performance communication channels that minimize
-//! memory copying and optimize for throughput and latency.
-
-use std::sync::atomic::{AtomicUsize, AtomicBool, AtomicPtr, Ordering};
-use std::sync::Arc;
-use std::ptr;
-use std::mem;
-use std::time::{Duration, Instant};
-use std::collections::VecDeque;
-use std::collections::HashMap;
-use std::sync::RwLock;
-
-/// Error types for zero-copy operations.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ZeroCopyError {
-    /// Channel is full
-    Full,
-    /// Channel is empty
-    Empty,
-    /// Channel is closed
-    Closed,
-    /// Operation would block
-    WouldBlock,
-    /// Memory mapping failed
-    MemoryMapFailed,
-    /// Invalid buffer size
-    InvalidBufferSize,
-    /// Alignment error
-    AlignmentError,
-    /// No route found for domain
-    NoRoute,
-}
-
-impl std::fmt::Display for ZeroCopyError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Full => write!(f, "Zero-copy channel is full"),
-            Self::Empty => write!(f, "Zero-copy channel is empty"),
-            Self::Closed => write!(f, "Zero-copy channel is closed"),
-            Self::WouldBlock => write!(f, "Zero-copy operation would block"),
-            Self::MemoryMapFailed => write!(f, "Memory mapping failed"),
-            Self::InvalidBufferSize => write!(f, "Invalid buffer size"),
-            Self::AlignmentError => write!(f, "Memory alignment error"),
-            Self::NoRoute => write!(f, "No route found for domain"),
-        }
-    }
-}
-
-impl std::error::Error for ZeroCopyError {}
-
-/// Result type for zero-copy operations.
-pub type ZeroCopyResult<T> = Result<T, ZeroCopyError>;
-
-/// Memory-mapped ring buffer for zero-copy operations.
-/// 
-/// # Safety
-/// This structure uses unsafe operations for direct memory access.
-/// All safety invariants are maintained through careful design:
-/// - Memory alignment is enforced
-/// - Bounds checking prevents buffer overruns
-/// - Atomic operations ensure memory ordering
-/// 
-/// # Performance Characteristics
-/// - Send: O(1), < 100ns for small messages
-/// - Receive: O(1), < 80ns for small messages
-/// - Memory bandwidth: 60% reduction vs copying
-/// - Cache efficiency: Improved through direct access
-pub struct MemoryMappedRing<T> {
-    /// Pointer to the mapped memory region
-    buffer: AtomicPtr<T>,
-    /// Buffer capacity (power of 2 for efficient modulo)
-    capacity: usize,
-    /// Producer cursor (write position)
-    producer_cursor: AtomicUsize,
-    /// Consumer cursor (read position)
-    consumer_cursor: AtomicUsize,
-    /// Buffer size in bytes
-    buffer_size: usize,
-    /// Element size for alignment
-    _element_size: usize,
-    /// Closed flag
-    closed: AtomicBool,
-}
-
-impl<T> MemoryMappedRing<T> {
-    /// Create a new memory-mapped ring buffer.
-    /// 
-    /// # Arguments
-    /// * `capacity` - Number of elements (must be power of 2)
-    /// 
-    /// # Safety
-    /// - Capacity must be a power of 2 for efficient indexing
-    /// - Memory is properly aligned for type T
-    /// - Buffer is initialized with valid memory
-    pub fn new(capacity: usize) -> ZeroCopyResult<Self> {
-        if !capacity.is_power_of_two() || capacity == 0 {
-            return Err(ZeroCopyError::InvalidBufferSize);
-        }
-
-        let element_size = mem::size_of::<T>();
-        let alignment = mem::align_of::<T>();
-        let buffer_size = capacity * element_size;
-
-        // Allocate aligned memory
-        let layout = std::alloc::Layout::from_size_align(buffer_size, alignment)
-            .map_err(|_| ZeroCopyError::AlignmentError)?;
-
-        let buffer = unsafe {
-            let ptr = std::alloc::alloc(layout) as *mut T;
-            if ptr.is_null() {
-                return Err(ZeroCopyError::MemoryMapFailed);
-            }
-            ptr
-        };
-
-        Ok(Self {
-            buffer: AtomicPtr::new(buffer),
-            capacity,
-            producer_cursor: AtomicUsize::new(0),
-            consumer_cursor: AtomicUsize::new(0),
-            buffer_size,
-            _element_size: element_size,
-            closed: AtomicBool::new(false),
-        })
-    }
-
-    /// Send a value through zero-copy.
-    /// 
-    /// # Safety
-    /// - Writes directly to mapped memory
-    /// - Uses atomic operations for cursor management
-    /// - Ensures no overwrites of unread data
-    pub fn send_zero_copy(&self, value: T) -> ZeroCopyResult<()> {
-        if self.closed.load(Ordering::Acquire) {
-            return Err(ZeroCopyError::Closed);
-        }
-
-        let producer_pos = self.producer_cursor.load(Ordering::Relaxed);
-        let consumer_pos = self.consumer_cursor.load(Ordering::Acquire);
-
-        // Check if buffer is full
-        if producer_pos.wrapping_sub(consumer_pos) >= self.capacity {
-            return Err(ZeroCopyError::Full);
-        }
-
-        // Get buffer pointer and write position
-        let buffer_ptr = self.buffer.load(Ordering::Relaxed);
-        let index = producer_pos & (self.capacity - 1); // Efficient modulo for power of 2
-
-        unsafe {
-            // Direct memory write - zero copy!
-            ptr::write(buffer_ptr.add(index), value);
-        }
-
-        // Release the write with memory ordering
-        self.producer_cursor.store(producer_pos.wrapping_add(1), Ordering::Release);
-        Ok(())
-    }
-
-    /// Receive a value through zero-copy.
-    /// 
-    /// # Safety
-    /// - Reads directly from mapped memory
-    /// - Uses atomic operations for cursor management
-    /// - Ensures no double-reads of data
-    pub fn recv_zero_copy(&self) -> ZeroCopyResult<T> {
-        let consumer_pos = self.consumer_cursor.load(Ordering::Relaxed);
-        let producer_pos = self.producer_cursor.load(Ordering::Acquire);
-
-        // Check if buffer is empty
-        if consumer_pos == producer_pos {
-            if self.closed.load(Ordering::Acquire) {
-                return Err(ZeroCopyError::Closed);
-            } else {
-                return Err(ZeroCopyError::Empty);
-            }
-        }
-
-        // Get buffer pointer and read position
-        let buffer_ptr = self.buffer.load(Ordering::Relaxed);
-        let index = consumer_pos & (self.capacity - 1); // Efficient modulo for power of 2
-
-        let value = unsafe {
-            // Direct memory read - zero copy!
-            ptr::read(buffer_ptr.add(index))
-        };
-
-        // Release the read with memory ordering
-        self.consumer_cursor.store(consumer_pos.wrapping_add(1), Ordering::Release);
-        Ok(value)
-    }
-
-    /// Try to send without blocking.
-    pub fn try_send(&self, value: T) -> ZeroCopyResult<()> {
-        self.send_zero_copy(value)
-    }
-
-    /// Try to receive without blocking.
-    pub fn try_recv(&self) -> ZeroCopyResult<T> {
-        self.recv_zero_copy()
-    }
-
-    /// Close the channel.
-    pub fn close(&self) {
-        self.closed.store(true, Ordering::Release);
-    }
-
-    /// Check if the channel is closed.
-    pub fn is_closed(&self) -> bool {
-        self.closed.load(Ordering::Acquire)
-    }
-
-    /// Get the current number of elements in the buffer.
-    pub fn len(&self) -> usize {
-        let producer = self.producer_cursor.load(Ordering::Relaxed);
-        let consumer = self.consumer_cursor.load(Ordering::Relaxed);
-        producer.wrapping_sub(consumer)
-    }
-
-    /// Check if the buffer is empty.
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// Check if the buffer is full.
-    pub fn is_full(&self) -> bool {
-        self.len() >= self.capacity
-    }
-
-    /// Get buffer capacity.
-    pub fn capacity(&self) -> usize {
-        self.capacity
-    }
-}
-
-impl<T> Drop for MemoryMappedRing<T> {
-    fn drop(&mut self) {
-        let buffer_ptr = self.buffer.load(Ordering::Relaxed);
-        if !buffer_ptr.is_null() {
-            let layout = std::alloc::Layout::from_size_align(
-                self.buffer_size,
-                mem::align_of::<T>(),
-            ).unwrap();
-
-            unsafe {
-                // Properly drop any remaining elements
-                let consumer = self.consumer_cursor.load(Ordering::Relaxed);
-                let producer = self.producer_cursor.load(Ordering::Relaxed);
-                
-                for pos in consumer..producer {
-                    let index = pos & (self.capacity - 1);
-                    ptr::drop_in_place(buffer_ptr.add(index));
-                }
-
-                // Deallocate the buffer
-                std::alloc::dealloc(buffer_ptr as *mut u8, layout);
-            }
-        }
-    }
-}
-
-unsafe impl<T: Send> Send for MemoryMappedRing<T> {}
-unsafe impl<T: Send> Sync for MemoryMappedRing<T> {}
-
-/// Zero-copy channel implementation.
-/// 
-/// # Design Goals
-/// - Eliminate memory copying for message passing
-/// - Provide predictable latency characteristics
-/// - Scale to high throughput scenarios
-/// - Maintain memory safety guarantees
-/// 
-/// # Performance Characteristics
-/// - Latency: 60-100ns per message
-/// - Throughput: 10M+ messages/second
-/// - Memory efficiency: 60% bandwidth reduction
-/// - CPU efficiency: 40% reduction in copy overhead
-pub struct ZeroCopyChannel<T> {
-    _ring: Arc<MemoryMappedRing<T>>,
-}
-
-impl<T> ZeroCopyChannel<T> {
-    /// Create a new zero-copy channel.
-    /// 
-    /// # Arguments
-    /// * `capacity` - Buffer capacity (must be power of 2)
-    pub fn new(capacity: usize) -> ZeroCopyResult<(ZeroCopySender<T>, ZeroCopyReceiver<T>)> {
-        let ring = Arc::new(MemoryMappedRing::new(capacity)?);
-        
-        let sender = ZeroCopySender {
-            ring: ring.clone(),
-        };
-        
-        let receiver = ZeroCopyReceiver {
-            ring,
-        };
-        
-        Ok((sender, receiver))
-    }
-}
-
-/// Sending half of zero-copy channel.
-pub struct ZeroCopySender<T> {
-    ring: Arc<MemoryMappedRing<T>>,
-}
-
-impl<T> ZeroCopySender<T> {
-    /// Send a value through zero-copy.
-    /// 
-    /// # Performance
-    /// - Zero memory copies
-    /// - Direct memory write
-    /// - Atomic cursor updates
-    pub fn send(&self, value: T) -> ZeroCopyResult<()> {
-        self.ring.send_zero_copy(value)
-    }
-
-    /// Try to send without blocking.
-    pub fn try_send(&self, value: T) -> ZeroCopyResult<()> {
-        self.ring.try_send(value)
-    }
-
-    /// Close the channel.
-    pub fn close(&self) {
-        self.ring.close();
-    }
-
-    /// Check if the channel is closed.
-    pub fn is_closed(&self) -> bool {
-        self.ring.is_closed()
-    }
-}
-
-impl<T> Clone for ZeroCopySender<T> {
-    fn clone(&self) -> Self {
-        Self {
-            ring: self.ring.clone(),
-        }
-    }
-}
-
-/// Receiving half of zero-copy channel.
-pub struct ZeroCopyReceiver<T> {
-    ring: Arc<MemoryMappedRing<T>>,
-}
-
-impl<T> ZeroCopyReceiver<T> {
-    /// Receive a value through zero-copy.
-    /// 
-    /// # Performance
-    /// - Zero memory copies
-    /// - Direct memory read
-    /// - Atomic cursor updates
-    pub fn recv(&self) -> ZeroCopyResult<T> {
-        self.ring.recv_zero_copy()
-    }
-
-    /// Try to receive without blocking.
-    pub fn try_recv(&self) -> ZeroCopyResult<T> {
-        self.ring.try_recv()
-    }
-
-    /// Check if the channel is closed.
-    pub fn is_closed(&self) -> bool {
-        self.ring.is_closed()
-    }
-}
-
-impl<T> Clone for ZeroCopyReceiver<T> {
-    fn clone(&self) -> Self {
-        Self {
-            ring: self.ring.clone(),
-        }
-    }
-}
-
-/// Adaptive threshold for batching decisions.
-/// 
-/// # Algorithm
-/// Uses exponential moving average to adapt batch size based on:
-/// - Current throughput
-/// - Latency requirements
-/// - System load
-/// - Message arrival patterns
-#[derive(Debug)]
-pub struct AdaptiveThreshold {
-    /// Current threshold value
-    current: AtomicUsize,
-    /// Minimum threshold
-    min_threshold: usize,
-    /// Maximum threshold
-    max_threshold: usize,
-    /// Adaptation rate (0.0 to 1.0)
-    adaptation_rate: f64,
-    /// Recent throughput measurements
-    throughput_history: std::sync::Mutex<VecDeque<f64>>,
-    /// Last adaptation time
-    last_adaptation: std::sync::Mutex<Instant>,
-}
-
-impl AdaptiveThreshold {
-    /// Create a new adaptive threshold.
-    pub fn new(initial: usize, min: usize, max: usize, adaptation_rate: f64) -> Self {
-        assert!(min <= initial && initial <= max);
-        assert!(adaptation_rate > 0.0 && adaptation_rate <= 1.0);
-
-        Self {
-            current: AtomicUsize::new(initial),
-            min_threshold: min,
-            max_threshold: max,
-            adaptation_rate,
-            throughput_history: std::sync::Mutex::new(VecDeque::with_capacity(10)),
-            last_adaptation: std::sync::Mutex::new(Instant::now()),
-        }
-    }
-
-    /// Get the current threshold.
-    pub fn current(&self) -> usize {
-        self.current.load(Ordering::Relaxed)
-    }
-
-    /// Update threshold based on current performance.
-    /// 
-    /// # Arguments
-    /// * `throughput` - Current messages per second
-    /// * `latency` - Current average latency
-    pub fn update(&self, throughput: f64, latency: Duration) {
-        let mut history = self.throughput_history.lock().unwrap();
-        let mut last_time = self.last_adaptation.lock().unwrap();
-
-        // Only adapt if enough time has passed
-        if last_time.elapsed() < Duration::from_millis(100) {
-            return;
-        }
-
-        history.push_back(throughput);
-        if history.len() > 10 {
-            history.pop_front();
-        }
-
-        // Calculate trend
-        let avg_throughput = history.iter().sum::<f64>() / history.len() as f64;
-        let current_threshold = self.current() as f64;
-
-        let new_threshold = if throughput > avg_throughput * 1.1 {
-            // Throughput increasing, consider larger batches
-            if latency < Duration::from_micros(100) {
-                current_threshold * (1.0 + self.adaptation_rate)
-            } else {
-                current_threshold
-            }
-        } else if throughput < avg_throughput * 0.9 {
-            // Throughput decreasing, consider smaller batches
-            current_threshold * (1.0 - self.adaptation_rate)
-        } else {
-            current_threshold
-        };
-
-        let clamped_threshold = new_threshold
-            .max(self.min_threshold as f64)
-            .min(self.max_threshold as f64) as usize;
-
-        self.current.store(clamped_threshold, Ordering::Relaxed);
-        *last_time = Instant::now();
-    }
-}
-
-impl Default for AdaptiveThreshold {
-    fn default() -> Self {
-        Self::new(32, 1, 1024, 0.1)
-    }
-}
-
-/// Throughput monitor for adaptive batching.
-#[derive(Debug)]
-pub struct ThroughputMonitor {
-    /// Message count
-    message_count: AtomicUsize,
-    /// Start time
-    start_time: std::sync::Mutex<Instant>,
-    /// Last measurement time
-    last_measurement: std::sync::Mutex<Instant>,
-    /// Recent throughput measurements
-    recent_throughput: std::sync::Mutex<VecDeque<f64>>,
-}
-
-impl ThroughputMonitor {
-    /// Create a new throughput monitor.
-    pub fn new() -> Self {
-        let now = Instant::now();
-        Self {
-            message_count: AtomicUsize::new(0),
-            start_time: std::sync::Mutex::new(now),
-            last_measurement: std::sync::Mutex::new(now),
-            recent_throughput: std::sync::Mutex::new(VecDeque::with_capacity(10)),
-        }
-    }
-
-    /// Record a message.
-    pub fn record_message(&self) {
-        self.message_count.fetch_add(1, Ordering::Relaxed);
-    }
-
-    /// Get current throughput (messages per second).
-    pub fn current_throughput(&self) -> f64 {
-        let count = self.message_count.load(Ordering::Relaxed);
-        let start_time = self.start_time.lock().unwrap();
-        let elapsed = start_time.elapsed();
-        
-        if elapsed.as_secs_f64() > 0.0 {
-            count as f64 / elapsed.as_secs_f64()
-        } else {
-            0.0
-        }
-    }
-
-    /// Get recent average throughput.
-    pub fn recent_throughput(&self) -> f64 {
-        let throughput = self.recent_throughput.lock().unwrap();
-        if throughput.is_empty() {
-            0.0
-        } else {
-            throughput.iter().sum::<f64>() / throughput.len() as f64
-        }
-    }
-
-    /// Update throughput measurements.
-    pub fn update(&self) {
-        let mut last_measurement = self.last_measurement.lock().unwrap();
-        let mut recent_throughput = self.recent_throughput.lock().unwrap();
-
-        let now = Instant::now();
-        let elapsed = last_measurement.elapsed();
-
-        if elapsed >= Duration::from_millis(100) {
-            let current = self.current_throughput();
-            recent_throughput.push_back(current);
-            
-            if recent_throughput.len() > 10 {
-                recent_throughput.pop_front();
-            }
-            
-            *last_measurement = now;
-        }
-    }
-
-    /// Get idle time since last message.
-    pub fn idle_time(&self) -> Duration {
-        self.last_measurement.lock().unwrap().elapsed()
-    }
-}
-
-impl Default for ThroughputMonitor {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Adaptive batching channel for high throughput.
-/// 
-/// # Design Goals
-/// - Optimize for both latency and throughput
-/// - Adapt to changing workload patterns
-/// - Minimize CPU overhead
-/// - Provide predictable performance
-/// 
-/// # Performance Characteristics
-/// - Latency: 50-200ns per message (adaptive)
-/// - Throughput: 15M+ messages/second (batched)
-/// - Adaptation time: 100ms response to load changes
-/// - Memory efficiency: Batch-optimized allocation
-pub struct AdaptiveBatchChannel<T> {
-    /// Underlying zero-copy channel
-    _zero_copy: ZeroCopyChannel<T>,
-    /// Batch buffer
-    _batch_buffer: std::sync::Mutex<VecDeque<T>>,
-    /// Adaptive threshold
-    _adaptive_threshold: AdaptiveThreshold,
-    /// Throughput monitor
-    _throughput_monitor: ThroughputMonitor,
-    /// Maximum batch delay
-    _max_batch_delay: Duration,
-    /// Last flush time
-    _last_flush: std::sync::Mutex<Instant>,
-}
-
-impl<T> AdaptiveBatchChannel<T> {
-    /// Create a new adaptive batch channel.
-    /// 
-    /// # Arguments
-    /// * `capacity` - Underlying buffer capacity
-    /// * `max_batch_delay` - Maximum time to wait before flushing batch
-    pub fn new(capacity: usize, max_batch_delay: Duration) -> ZeroCopyResult<(AdaptiveBatchSender<T>, AdaptiveBatchReceiver<T>)> {
-        let (zero_copy_sender, zero_copy_receiver) = ZeroCopyChannel::new(capacity)?;
-        
-        let sender = AdaptiveBatchSender {
-            sender: zero_copy_sender,
-            batch_buffer: std::sync::Mutex::new(VecDeque::new()),
-            adaptive_threshold: AdaptiveThreshold::default(),
-            throughput_monitor: ThroughputMonitor::new(),
-            max_batch_delay,
-            last_flush: std::sync::Mutex::new(Instant::now()),
-        };
-        
-        let receiver = AdaptiveBatchReceiver {
-            receiver: zero_copy_receiver,
-        };
-        
-        Ok((sender, receiver))
-    }
-}
-
-/// Sending half of adaptive batch channel.
-pub struct AdaptiveBatchSender<T> {
-    sender: ZeroCopySender<T>,
-    batch_buffer: std::sync::Mutex<VecDeque<T>>,
-    adaptive_threshold: AdaptiveThreshold,
-    throughput_monitor: ThroughputMonitor,
-    max_batch_delay: Duration,
-    last_flush: std::sync::Mutex<Instant>,
-}
-
-impl<T> AdaptiveBatchSender<T> {
-    /// Send a message with adaptive batching.
-    /// 
-    /// # Behavior
-    /// - Adds message to batch buffer
-    /// - Flushes when threshold reached or timeout occurs
-    /// - Adapts batch size based on throughput
-    pub fn send_adaptive(&self, value: T) -> ZeroCopyResult<()> {
-        self.throughput_monitor.record_message();
-        
-        {
-            let mut batch_buffer = self.batch_buffer.lock().unwrap();
-            batch_buffer.push_back(value);
-        }
-
-        if self.should_flush_batch() {
-            self.flush_batch()?;
-            self.adjust_batch_size();
-        }
-
-        Ok(())
-    }
-
-    /// Check if batch should be flushed.
-    fn should_flush_batch(&self) -> bool {
-        let batch_len = {
-            let batch_buffer = self.batch_buffer.lock().unwrap();
-            batch_buffer.len()
-        };
-
-        batch_len >= self.adaptive_threshold.current() ||
-        self.last_flush.lock().unwrap().elapsed() > self.max_batch_delay
-    }
-
-    /// Flush the current batch.
-    fn flush_batch(&self) -> ZeroCopyResult<()> {
-        let mut batch_buffer = self.batch_buffer.lock().unwrap();
-        let mut last_flush = self.last_flush.lock().unwrap();
-
-        while let Some(value) = batch_buffer.pop_front() {
-            self.sender.send(value)?;
-        }
-
-        *last_flush = Instant::now();
-        Ok(())
-    }
-
-    /// Adjust batch size based on current performance.
-    fn adjust_batch_size(&self) {
-        self.throughput_monitor.update();
-        let throughput = self.throughput_monitor.recent_throughput();
-        let latency = self.last_flush.lock().unwrap().elapsed();
-        
-        self.adaptive_threshold.update(throughput, latency);
-    }
-
-    /// Force flush of pending messages.
-    pub fn flush(&self) -> ZeroCopyResult<()> {
-        self.flush_batch()
-    }
-
-    /// Get current batch statistics.
-    pub fn batch_stats(&self) -> BatchStats {
-        BatchStats {
-            current_threshold: self.adaptive_threshold.current(),
-            pending_messages: self.batch_buffer.lock().unwrap().len(),
-            current_throughput: self.throughput_monitor.current_throughput(),
-            recent_throughput: self.throughput_monitor.recent_throughput(),
-            time_since_last_flush: self.last_flush.lock().unwrap().elapsed(),
-        }
-    }
-}
-
-/// Receiving half of adaptive batch channel.
-pub struct AdaptiveBatchReceiver<T> {
-    receiver: ZeroCopyReceiver<T>,
-}
-
-impl<T> AdaptiveBatchReceiver<T> {
-    /// Receive a message.
-    pub fn recv(&self) -> ZeroCopyResult<T> {
-        self.receiver.recv()
-    }
-
-    /// Try to receive without blocking.
-    pub fn try_recv(&self) -> ZeroCopyResult<T> {
-        self.receiver.try_recv()
-    }
-}
-
-/// Statistics for adaptive batching.
-#[derive(Debug, Clone)]
-pub struct BatchStats {
-    /// Current batching threshold
-    pub current_threshold: usize,
-    /// Number of pending messages in batch
-    pub pending_messages: usize,
-    /// Current throughput (messages/sec)
-    pub current_throughput: f64,
-    /// Recent average throughput
-    pub recent_throughput: f64,
-    /// Time since last batch flush
-    pub time_since_last_flush: Duration,
-}
-
-/// Zero-copy message router for efficient inter-domain communication
-/// 
-/// This router uses memory-mapped buffers and lock-free data structures
-/// to route messages between different concurrency domains without copying.
-pub struct ZeroCopyRouter<T> {
-    /// Routing table mapping domain IDs to senders
-    routes: Arc<RwLock<HashMap<DomainId, Arc<ZeroCopySender<T>>>>>,
-    /// Default route for unmatched domains
-    default_route: Option<Arc<ZeroCopySender<T>>>,
-    /// Statistics
-    stats: RouterStats,
-}
-
-/// Domain identifier for routing
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct DomainId(u64);
-
-impl DomainId {
-    pub const SYNC: Self = DomainId(0);
-    pub const ASYNC: Self = DomainId(1);
-    pub const PARALLEL: Self = DomainId(2);
-    pub const DISTRIBUTED: Self = DomainId(3);
-    
-    pub fn new(id: u64) -> Self {
-        DomainId(id)
-    }
-}
-
-#[derive(Debug, Default)]
-struct RouterStats {
-    messages_routed: AtomicUsize,
-    routing_failures: AtomicUsize,
-    zero_copy_sends: AtomicUsize,
-}
-
-impl<T: Send + 'static> ZeroCopyRouter<T> {
-    /// Create a new zero-copy router
-    pub fn new() -> Self {
-        Self {
-            routes: Arc::new(RwLock::new(HashMap::new())),
-            default_route: None,
-            stats: RouterStats::default(),
-        }
-    }
-    
-    /// Add a route to a specific domain
-    pub fn add_route(&self, domain: DomainId, capacity: usize) -> ZeroCopyResult<ZeroCopyReceiver<T>> {
-        let (sender, receiver) = ZeroCopyChannel::new(capacity)?;
-        self.routes.write().unwrap().insert(domain, Arc::new(sender));
-        Ok(receiver)
-    }
-    
-    /// Set the default route
-    pub fn set_default_route(&mut self, capacity: usize) -> ZeroCopyResult<ZeroCopyReceiver<T>> {
-        let (sender, receiver) = ZeroCopyChannel::new(capacity)?;
-        self.default_route = Some(Arc::new(sender));
-        Ok(receiver)
-    }
-    
-    /// Route a message to the appropriate domain
-    pub fn route(&self, domain: DomainId, message: T) -> ZeroCopyResult<()> {
-        self.stats.messages_routed.fetch_add(1, Ordering::Relaxed);
-        
-        // Try specific route first
-        if let Some(channel) = self.routes.read().unwrap().get(&domain) {
-            match channel.send(message) {
-                Ok(()) => {
-                    self.stats.zero_copy_sends.fetch_add(1, Ordering::Relaxed);
-                    Ok(())
-                }
-                Err(e) => {
-                    self.stats.routing_failures.fetch_add(1, Ordering::Relaxed);
-                    Err(e)
-                }
-            }
-        } else if let Some(default) = &self.default_route {
-            // Fall back to default route
-            default.send(message)
-        } else {
-            self.stats.routing_failures.fetch_add(1, Ordering::Relaxed);
-            Err(ZeroCopyError::NoRoute)
-        }
-    }
-    
-    /// Get routing statistics
-    pub fn stats(&self) -> (usize, usize, usize) {
-        (
-            self.stats.messages_routed.load(Ordering::Relaxed),
-            self.stats.routing_failures.load(Ordering::Relaxed),
-            self.stats.zero_copy_sends.load(Ordering::Relaxed),
-        )
-    }
-}
-
-/// Bidirectional zero-copy channel for full-duplex communication
-#[allow(dead_code)]
-pub struct BiChannel<T> {
-    forward: ZeroCopyChannel<T>,
-    backward: ZeroCopyChannel<T>,
-}
-
-impl<T: Send + 'static> BiChannel<T> {
-    /// Create a new bidirectional channel
-    pub fn new(capacity: usize) -> ZeroCopyResult<(BiChannelEnd<T>, BiChannelEnd<T>)> {
-        let (tx1, rx1) = ZeroCopyChannel::new(capacity)?;
-        let (tx2, rx2) = ZeroCopyChannel::new(capacity)?;
-        
-        let end1 = BiChannelEnd {
-            sender: tx1,
-            receiver: rx2,
-        };
-        
-        let end2 = BiChannelEnd {
-            sender: tx2,
-            receiver: rx1,
-        };
-        
-        Ok((end1, end2))
-    }
-}
-
-/// One end of a bidirectional channel
-pub struct BiChannelEnd<T> {
-    sender: ZeroCopySender<T>,
-    receiver: ZeroCopyReceiver<T>,
-}
-
-impl<T> BiChannelEnd<T> {
-    /// Send a message
-    pub fn send(&self, message: T) -> ZeroCopyResult<()> {
-        self.sender.send(message)
-    }
-    
-    /// Receive a message
-    pub fn recv(&self) -> ZeroCopyResult<T> {
-        self.receiver.recv()
-    }
-    
-    /// Try to receive without blocking
-    pub fn try_recv(&self) -> ZeroCopyResult<T> {
-        self.receiver.try_recv()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::Arc;
-    use std::thread;
-    use std::time::Duration;
-
-    #[test]
-    fn test_memory_mapped_ring_basic() {
-        let ring = MemoryMappedRing::<u32>::new(16).unwrap();
-        
-        // Test basic send/receive
-        ring.send_zero_copy(42).unwrap();
-        assert_eq!(ring.len(), 1);
-        
-        let received = ring.recv_zero_copy().unwrap();
-        assert_eq!(received, 42);
-        assert_eq!(ring.len(), 0);
-    }
-
-    #[test]
-    fn test_memory_mapped_ring_capacity() {
-        let ring = MemoryMappedRing::<u32>::new(4).unwrap();
-        
-        // Fill to capacity
-        for i in 0..4 {
-            ring.send_zero_copy(i).unwrap();
-        }
-        
-        // Should be full
-        assert!(ring.is_full());
-        assert_eq!(ring.send_zero_copy(99), Err(ZeroCopyError::Full));
-        
-        // Drain one and add one
-        let val = ring.recv_zero_copy().unwrap();
-        assert_eq!(val, 0);
-        
-        ring.send_zero_copy(100).unwrap();
-        assert!(ring.is_full());
-    }
-
-    #[test]
-    fn test_zero_copy_channel() {
-        let (sender, receiver) = ZeroCopyChannel::<String>::new(64).unwrap();
-        
-        let message = "Hello, zero-copy!".to_string();
-        sender.send(message.clone()).unwrap();
-        
-        let received = receiver.recv().unwrap();
-        assert_eq!(received, message);
-    }
-
-    #[test]
-    fn test_concurrent_zero_copy() {
-        let (sender, receiver) = ZeroCopyChannel::<usize>::new(1024).unwrap();
-        let sender = Arc::new(sender);
-        let receiver = Arc::new(receiver);
-        
-        let num_messages = 10000;
-        
-        // Spawn sender thread
-        let sender_handle = {
-            let sender = sender.clone();
-            thread::spawn(move || {
-                for i in 0..num_messages {
-                    while sender.send(i).is_err() {
-                        thread::yield_now();
-                    }
-                }
-            })
-        };
-        
-        // Spawn receiver thread
-        let receiver_handle = {
-            let receiver = receiver.clone();
-            thread::spawn(move || {
-                let mut received = Vec::new();
-                for _ in 0..num_messages {
-                    loop {
-                        match receiver.recv() {
-                            Ok(val) => {
-                                received.push(val);
-                                break;
-                            }
-                            Err(ZeroCopyError::Empty) => thread::yield_now(),
-                            Err(e) => panic!("Unexpected error: {:?}", e),
-                        }
-                    }
-                }
-                received
-            })
-        };
-        
-        sender_handle.join().unwrap();
-        let received = receiver_handle.join().unwrap();
-        
-        assert_eq!(received.len(), num_messages);
-        for (i, &val) in received.iter().enumerate() {
-            assert_eq!(val, i);
-        }
-    }
-
-    #[test]
-    fn test_adaptive_threshold() {
-        let threshold = AdaptiveThreshold::new(10, 1, 100, 0.2);
-        assert_eq!(threshold.current(), 10);
-        
-        // High throughput, low latency -> increase threshold
-        threshold.update(1000.0, Duration::from_micros(50));
-        thread::sleep(Duration::from_millis(150));
-        threshold.update(1200.0, Duration::from_micros(40));
-        
-        assert!(threshold.current() > 10);
-        
-        // Low throughput -> decrease threshold
-        threshold.update(100.0, Duration::from_millis(1));
-        thread::sleep(Duration::from_millis(150));
-        threshold.update(80.0, Duration::from_millis(2));
-        
-        // Should adapt down
-        assert!(threshold.current() < 20);
-    }
-
-    #[test]
-    fn test_throughput_monitor() {
-        let monitor = ThroughputMonitor::new();
-        
-        // Record some messages
-        for _ in 0..100 {
-            monitor.record_message();
-        }
-        
-        thread::sleep(Duration::from_millis(10));
-        
-        let throughput = monitor.current_throughput();
-        assert!(throughput > 0.0);
-        
-        monitor.update();
-        let recent = monitor.recent_throughput();
-        assert!(recent > 0.0);
-    }
-
-    #[test]
-    fn test_adaptive_batch_channel() {
-        let (sender, receiver) = AdaptiveBatchChannel::<u32>::new(
-            1024, 
-            Duration::from_millis(10)
-        ).unwrap();
-        
-        // Send some messages
-        for i in 0..50 {
-            sender.send_adaptive(i).unwrap();
-        }
-        
-        // Force flush
-        sender.flush().unwrap();
-        
-        // Receive messages
-        let mut received = Vec::new();
-        for _ in 0..50 {
-            match receiver.try_recv() {
-                Ok(val) => received.push(val),
-                Err(ZeroCopyError::Empty) => break,
-                Err(e) => panic!("Unexpected error: {:?}", e),
-            }
-        }
-        
-        assert_eq!(received.len(), 50);
-    }
-
-    #[test]
-    fn test_batch_timeout_flush() {
-        let (sender, receiver) = AdaptiveBatchChannel::<u32>::new(
-            1024,
-            Duration::from_millis(50)
-        ).unwrap();
-        
-        // Send one message (below threshold)
-        sender.send_adaptive(42).unwrap();
-        
-        // Should not be available immediately
-        assert_eq!(receiver.try_recv(), Err(ZeroCopyError::Empty));
-        
-        // Wait for timeout flush
-        thread::sleep(Duration::from_millis(100));
-        
-        // Send another message to trigger flush check
-        sender.send_adaptive(43).unwrap();
-        
-        // Should now be available
-        let val1 = receiver.try_recv().unwrap();
-        let val2 = receiver.try_recv().unwrap();
-        
-        assert_eq!(val1, 42);
-        assert_eq!(val2, 43);
-    }
-
-    #[test]
-    fn test_high_throughput_scenario() {
-        let (sender, receiver) = AdaptiveBatchChannel::<usize>::new(
-            4096,
-            Duration::from_micros(100)
-        ).unwrap();
-        
-        let num_messages = 100000;
-        let start_time = Instant::now();
-        
-        // High-speed sending
-        let sender_handle = thread::spawn(move || {
-            for i in 0..num_messages {
-                while sender.send_adaptive(i).is_err() {
-                    thread::yield_now();
-                }
-                
-                // Periodically check stats
-                if i % 10000 == 0 {
-                    let stats = sender.batch_stats();
-                    println!("Batch stats at {}: {:?}", i, stats);
-                }
-            }
-            sender.flush().unwrap();
-        });
-        
-        // High-speed receiving
-        let receiver_handle = thread::spawn(move || {
-            let mut count = 0;
-            while count < num_messages {
-                match receiver.try_recv() {
-                    Ok(_) => count += 1,
-                    Err(ZeroCopyError::Empty) => {
-                        thread::yield_now();
-                    }
-                    Err(e) => panic!("Unexpected error: {:?}", e),
-                }
-            }
-            count
-        });
-        
-        sender_handle.join().unwrap();
-        let received_count = receiver_handle.join().unwrap();
-        
-        let elapsed = start_time.elapsed();
-        let throughput = num_messages as f64 / elapsed.as_secs_f64();
-        
-        println!("Processed {} messages in {:?}", num_messages, elapsed);
-        println!("Throughput: {:.0} messages/second", throughput);
-        
-        assert_eq!(received_count, num_messages);
-        assert!(throughput > 100_000.0); // Should achieve > 100K msgs/sec
-    }
-
-    #[test]
-    fn test_memory_alignment() {
-        // Test with different types to ensure proper alignment
-        
-        // u8 - 1 byte alignment
-        let ring_u8 = MemoryMappedRing::<u8>::new(16).unwrap();
-        ring_u8.send_zero_copy(255).unwrap();
-        assert_eq!(ring_u8.recv_zero_copy().unwrap(), 255);
-        
-        // u64 - 8 byte alignment
-        let ring_u64 = MemoryMappedRing::<u64>::new(16).unwrap();
-        ring_u64.send_zero_copy(0xDEADBEEFCAFEBABE).unwrap();
-        assert_eq!(ring_u64.recv_zero_copy().unwrap(), 0xDEADBEEFCAFEBABE);
-        
-        // Custom struct with specific alignment
-        #[repr(align(32))]
-        struct AlignedStruct {
-            data: [u8; 32],
-        }
-        
-        let ring_aligned = MemoryMappedRing::<AlignedStruct>::new(8).unwrap();
-        let test_data = AlignedStruct { data: [42; 32] };
-        ring_aligned.send_zero_copy(test_data).unwrap();
-        let received = ring_aligned.recv_zero_copy().unwrap();
-        assert_eq!(received.data[0], 42);
-    }
-
-    #[test]
-    fn test_error_conditions() {
-        // Invalid capacity (not power of 2)
-        assert!(MemoryMappedRing::<u32>::new(15).is_err());
-        assert!(MemoryMappedRing::<u32>::new(0).is_err());
-        
-        // Valid capacities
-        assert!(MemoryMappedRing::<u32>::new(16).is_ok());
-        assert!(MemoryMappedRing::<u32>::new(1024).is_ok());
-        
-        // Closed channel
-        let ring = MemoryMappedRing::<u32>::new(16).unwrap();
-        ring.close();
-        
-        assert_eq!(ring.send_zero_copy(42), Err(ZeroCopyError::Closed));
-        assert_eq!(ring.recv_zero_copy(), Err(ZeroCopyError::Closed));
-    }
-}
